+++ source/packages/services/device-patcher/src/utils/errors.ts
@@ -1,4 +1,4 @@
-/*********************************************************************************************************************
+/********************************************************************************************************************
  *  Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.                                           *
  *                                                                                                                    *
  *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
@@ -21,7 +21,8 @@ export function handleError(e:Error, res:Response): void {
         e.message.startsWith('MISSING_REQUIRED') ||
         e.message === 'FAILED_VALIDATION' ||
         e.message === 'UNSUPPORTED_TRANSITION' ||
-        e.message === 'UNSUPPORTED_DEPLOYMENT_STATUS'
+        e.message === 'UNSUPPORTED_PATCH_STATUS' ||
+        e.message === 'BAD_REQUEST'
     ) {
         res.status(400).json({error: e.message}).end();
     } else if (
@@ -34,7 +35,7 @@ export function handleError(e:Error, res:Response): void {
         res.status(404).json({error: e.message}).end();
     } else if (
         e.name === 'ConditionalCheckFailedException' ||
-        e.message.indexOf('with id already exists')>=0 )
+        e.message === 'CONFLICT')
     {
         res.status(409).json({error: 'Item already exists'}).end();
     } else if (