+++ source/packages/services/device-patcher/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @cdf/device-patcher
 
-This log was last generated on Mon, 28 Mar 2022 00:28:46 GMT and should not be manually modified.
+This log was last generated on Thu, 12 May 2022 21:12:39 GMT and should not be manually modified.
+
+## 1.0.0
+Thu, 12 May 2022 21:12:39 GMT
+
+### Breaking changes
+
+- unbreak, state manager association (ansible execution) run documents and improve Patch templates API 
 
 ## 0.1.2
 Mon, 28 Mar 2022 00:28:46 GMT