+++ source/packages/services/installer/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @cdf/installer
 
-This log was last generated on Thu, 12 May 2022 01:58:21 GMT and should not be manually modified.
+This log was last generated on Thu, 12 May 2022 21:12:39 GMT and should not be manually modified.
+
+## 0.4.2
+Thu, 12 May 2022 21:12:39 GMT
+
+### Patches
+
+- minor updates for device-patcher deployment
+
+## 0.4.1
+Thu, 12 May 2022 03:51:07 GMT
+
+### Patches
+
+- minor bug fix, to fix issue where installer would fail due to empty answers
 
 ## 0.4.0
 Thu, 12 May 2022 01:58:21 GMT